requirements.txt
flask
gunicorn
tabula-py
pandas
boto3==1.3.1
Jinja2==2.8
Werkzeug
